+++ packages/lemon_js/lib/lemon_js.dart
@@ -18,4 +18,5 @@ export 'src/runtime/quickjs_value.dart';
 export 'src/module/quickjs_web_crypto_mount.dart';
 export 'src/module/quickjs_fetch_mount.dart';
 export 'src/module/quickjs_axios_mount.dart';
+export 'src/module/quickjs_key_value_storage.dart';
 export 'src/module/quickjs_websocket_mount.dart';